Plug and Play SD Card to Music Playback
With the current code, insert an SD Card with MP3 files at the root directory. The code should automatically starting playing in alphabetically sorted order, each song only playing for a short while before moving to the next.

Note that sometimes, when you are copying MP3 files into your SD card from a computer, the write process may not be fully successful, creating corrupted files that the code is unable to playback.