Noise Testing
There is a constant but quiet "old TV" static noise using headphones (and probably if you tested speaker output too). This exists even while you have I2S on or off. It also persists when the ESP32s3 is off, and the only way to remove the static is to powerdown the output using registers or power down the entire amp WM8731. There's a high change that the headphone amplifier is increasing the "floor noise" generated by a noisy power supply (because we are using a switching regulator). Perhaps using a modern amplifier like the MAX9850 would be fine? This is different from the clicking sounds generated while playing mp3 or wav.