Buzzing and Beeping from SD Card
Recall that MP3 and WAV playback were producing horrible ticking noises and buzzing noises respectively. I first targeted WAV playback as it was a much simpler code loop. Increasing the buffer size seemed to decrease the pitch of the noise and it's frequency (how many times it ticked per second). This introduced to possible factors - either the file read from the SD Card was responsible for this change or the file write to the I2S was. To isolate the issue, I got rid of SD Card initialization and uploaded a small WAV file into a VFS partition in the ESP flash using SPIFFS. I was able to play audio through I2S without using SD Card, and there little to no noise received from the headphone output. Next, I then created a thread that constantly read a file from the SD Card while I2S was still using SPIFFS. Despite I2S not using SD Card, the SD Card thread that was file reading was somehow producing noise that was coupled into the headphone output, whether I2S was active or not. Adding a delay to the SD Card reading thread decreased the pitch and frequency (and volume) of the noise, and these changes were very noticeable.

In short, probably because of poor ground design on the board, the SD Card data transaction produced noises that were absorbed and outputted by the headphone jack. This then means that the reason why MP3 playback had less noise than WAV playback is because MP3 data is compressed data, meaning each read produces more audio output than a WAV does, which is 1:1. Thus, to produce the same bitrate, the code needs to read less. Additionally, each loop time is elongated due to decompression, then added on top of I2S blocking, the frequency at which the code was reading the SD Card was lowered, thus producing less noise.

This also raises an implication that the background static, aside from the buzzing, is probably a hardware problem too. It could be tied to the buck converter which uses high frequency switching for power regulation, often known to generate noise on the power line.